Move mixins and utils from components dir
