+++ src/client/components/Board.vue
@@ -156,7 +156,7 @@ import {IAresData} from '@/ares/IAresData';
 import {SpaceModel} from '@/models/SpaceModel';
 import {SpaceType} from '@/SpaceType';
 import {SpaceId} from '@/boards/ISpace';
-import {TranslateMixin} from '@/client/components/TranslateMixin';
+import {TranslateMixin} from '@/client/mixins/TranslateMixin';
 
 class GlobalParamLevel {
   constructor(public value: number, public isActive: boolean, public strValue: string) {
