+++ src/client/components/GlobalParameterValue.vue
@@ -15,7 +15,7 @@
 <script lang="ts">
 
 import Vue from 'vue';
-import {TranslateMixin} from '@/client/components/TranslateMixin';
+import {TranslateMixin} from '@/client/mixins/TranslateMixin';
 import {MAX_OCEAN_TILES, MAX_OXYGEN_LEVEL, MAX_TEMPERATURE, MAX_VENUS_SCALE} from '@/constants';
 import {GlobalParameter} from '@/GlobalParameter';
