+++ src/client/components/LogPanel.vue
@@ -48,8 +48,8 @@ import {CardName} from '@/CardName';
 import {TileType} from '@/TileType';
 import {playerColorClass} from '@/utils/utils';
 import {Color} from '@/Color';
-import {SoundManager} from '@/client/components/SoundManager';
-import {PreferencesManager} from '@/client/components/PreferencesManager';
+import {SoundManager} from '@/client/utils/SoundManager';
+import {PreferencesManager} from '@/client/utils/PreferencesManager';
 import {GlobalEventName} from '@/turmoil/globalEvents/GlobalEventName';
 import GlobalEvent from '@/client/components/GlobalEvent.vue';
 import {getGlobalEventByName} from '@/turmoil/globalEvents/GlobalEventDealer';
