+++ src/client/components/OtherPlayer.vue
@@ -3,7 +3,7 @@
 import Vue from 'vue';
 
 import StackedCards from '@/client/components/StackedCards.vue';
-import {PlayerMixin} from '@/client/components/PlayerMixin';
+import {PlayerMixin} from '@/client/mixins/PlayerMixin';
 import {PublicPlayerModel} from '@/models/PlayerModel';
 import {mainAppSettings} from '@/client/components/App';
 import Card from '@/client/components/card/Card.vue';
