+++ src/client/components/PlayerHome.vue
@@ -238,13 +238,13 @@ import WaitingFor from '@/client/components/WaitingFor.vue';
 import Sidebar from '@/client/components/Sidebar.vue';
 import Colony from '@/client/components/Colony.vue';
 import LogPanel from '@/client/components/LogPanel.vue';
-import {PlayerMixin} from '@/client/components/PlayerMixin';
+import {PlayerMixin} from '@/client/mixins/PlayerMixin';
 import Turmoil from '@/client/components/Turmoil.vue';
 import {playerColorClass} from '@/utils/utils';
 import DynamicTitle from '@/client/components/common/DynamicTitle.vue';
 import SortableCards from '@/client/components/SortableCards.vue';
 import TopBar from '@/client/components/TopBar.vue';
-import {PreferencesManager} from '@/client/components/PreferencesManager';
+import {PreferencesManager} from '@/client/utils/PreferencesManager';
 import {KeyboardNavigation} from '@/KeyboardNavigation';
 import MoonBoard from '@/client/components/moon/MoonBoard.vue';
 import {Phase} from '@/Phase';
