+++ src/client/components/SelectPartyToSendDelegate.vue
@@ -17,7 +17,7 @@ import Vue from 'vue';
 import Button from '@/client/components/common/Button.vue';
 import {PlayerInputModel} from '@/models/PlayerInputModel';
 import Party from '@/client/components/Party.vue';
-import {TranslateMixin} from '@/client/components/TranslateMixin';
+import {TranslateMixin} from '@/client/mixins/TranslateMixin';
 import {PartyName} from '@/turmoil/parties/PartyName';
 
 export default Vue.extend({
