+++ src/client/components/card/CardNumber.vue
@@ -5,7 +5,7 @@
 <script lang="ts">
 
 import Vue from 'vue';
-import {PreferencesManager} from '@/client/components/PreferencesManager';
+import {PreferencesManager} from '@/client/utils/PreferencesManager';
 
 export default Vue.extend({
   name: 'CardNumber',
