+++ src/client/components/common/ConfirmDialog.vue
@@ -1,6 +1,6 @@
 <script lang="ts">
 import Vue from 'vue';
-import {TranslateMixin} from '@/client/components/TranslateMixin';
+import {TranslateMixin} from '@/client/mixins/TranslateMixin';
 
 const dialogPolyfill = require('dialog-polyfill');
