+++ src/client/components/create/CardsFilter.vue
@@ -27,7 +27,7 @@
 import Vue from 'vue';
 import {CardName} from '@/CardName';
 import {ALL_PRELUDE_CARD_NAMES, ALL_PROJECT_CARD_NAMES} from '@/cards/AllCards';
-import {TranslateMixin} from '@/client/components/TranslateMixin';
+import {TranslateMixin} from '@/client/mixins/TranslateMixin';
 import Button from '@/client/components/common/Button.vue';
 
 const allItems: Array<CardName> = ALL_PROJECT_CARD_NAMES.concat(ALL_PRELUDE_CARD_NAMES).sort();
