+++ src/client/components/overview/PlayerResource.vue
@@ -17,7 +17,7 @@
 import Vue from 'vue';
 import {DEFAULT_STEEL_VALUE, DEFAULT_TITANIUM_VALUE} from '@/constants';
 import {Resources} from '@/Resources';
-import {PreferencesManager} from '@/client/components/PreferencesManager';
+import {PreferencesManager} from '@/client/utils/PreferencesManager';
 
 export default Vue.extend({
   name: 'PlayerResource',
