+++ src/client/utils/PreferencesManager.ts
