+++ tests/components/PaymentTester.ts
@@ -1,7 +1,7 @@
 import Vue from 'vue';
 import {Wrapper} from '@vue/test-utils';
 import {expect} from 'chai';
-import {SelectHowToPayModel} from '@/client/components/PaymentWidgetMixin';
+import {SelectHowToPayModel} from '@/client/mixins/PaymentWidgetMixin';
 
 export type Unit = 'heat' | 'steel' | 'titanium' | 'floaters' | 'microbes' | 'megaCredits' | 'science';
